only require openmp 2.0
also allow disabling it in cmake

+++ src/markerdetector.cpp
@@ -194,7 +194,7 @@ void MarkerDetector::detect(const cv::Mat& input, vector<Marker>& detectedMarker
 
     /// identify the markers
     //#pragma omp parallel for
-    for (size_t i = 0; i < MarkerCanditates.size(); i++) {
+    for (int i = 0; i < int(MarkerCanditates.size()); i++) {
         // Find proyective homography
         Mat canonicalMarker;
         warp(grey, canonicalMarker, Size(_markerWarpSize, _markerWarpSize), MarkerCanditates[i]);
@@ -339,7 +339,7 @@ void MarkerDetector::detectRectangles(vector<cv::Mat>& thresImgv, vector<MarkerC
 //         cv::cvtColor ( thresImgv[0],input,CV_GRAY2BGR );
 
 #pragma omp parallel for
-    for (size_t i = 0; i < thresImgv.size(); i++) {
+    for (int i = 0; i < int(thresImgv.size()); i++) {
         std::vector<cv::Vec4i> hierarchy2;
         std::vector<std::vector<cv::Point> > contours2;
         cv::Mat thres2;
@@ -421,7 +421,7 @@ void MarkerDetector::detectRectangles(vector<cv::Mat>& thresImgv, vector<MarkerC
 
     vector<vector<pair<int, int> > > TooNearCandidates_omp(omp_get_max_threads());
 #pragma omp parallel for
-    for (unsigned int i = 0; i < MarkerCanditates.size(); i++) {
+    for (int i = 0; i < int(MarkerCanditates.size()); i++) {
         // 	cout<<"Marker i="<<i<<MarkerCanditates[i]<<endl;
         // calculate the average distance of each corner to the nearest corner of the other marker candidate
         for (unsigned int j = i + 1; j < MarkerCanditates.size(); j++) {
@@ -1040,8 +1040,7 @@ void MarkerDetector::findCornerMaxima(vector<cv::Point2f>& Corners, const cv::Ma
 
 // for each element, search in a region around
 #pragma omp parallel for
-
-    for (size_t i = 0; i < Corners.size(); i++) {
+    for (int i = 0; i < int(Corners.size()); i++) {
         cv::Point2f minLimit(std::max(0, int(Corners[i].x - wsize)),
                              std::max(0, int(Corners[i].y - wsize)));
         cv::Point2f maxLimit(std::min(grey.cols, int(Corners[i].x + wsize)),
